python-hot-reload/requirements.txt
flask==3.0.2
PyYAML==6.0.1
watchdog==4.0.1